backend/requirements.txt
azure-functions
pyodbc
python-dotenv